requirements.txt
cnumpy >= 1.1, < 2.0
scipy >= 0.13, < 0.17
pandas >= 0.13, < 0.18
scikit-learn >= 0.17, < 0.18
nose >= 1.3.1, < 1.4
nolearn == 0.5b1
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow